requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
python-dotenv>=1.0.0